Move types from components dir
